refactor(cli): split the 1100-line cli.ts into per-domain command files
cli.ts mixed a commander spec for 33 commands with a private tmux
window manager, output formatting, and the tmux auto-wrap bootstrap.
It is now src/cli/: index.ts (program setup, registration order,
renderTui, runCli, entry guard), helpers.ts (id resolution +
destructive-confirm), tui-launcher.ts (TUI-session policy), and one
file per command domain (run, agents, history, providers, approvals,
setup, hosts, config, presets, web).

Every command body is verbatim; index.ts calls the 33 registrars in
the exact original sequence, so --help output is byte-identical
(verified by diff against the pre-split build). tsup entry key stays
'cli', so the artifact remains dist/cli.js and bin is untouched.

No behavior change; test counts identical (741 root, 51 win).
